completed refactoring for storing the models and also providing with ability to reuse
