+++ src/Bet.Extensions.ML/README.md
@@ -1,0 +1,11 @@
+﻿# Bet.Extensions.ML
+
+A library that provides production ready extensions to [ML.NET library](https://github.com/dotnet/machinelearning).
+It contains two major functionality:
+
+1. ML.NET ModelBuilder
+- ML.NET models builder based on ML.NET AutoML generation idea
+
+2. ML.NET Web Api hosting
+- Utilizing `ObjectPool` similar to  [Extensions.ML](https://github.com/glennc/Extensions.ML)
+
